Make counter uint16

+++ client.go
@@ -65,7 +65,7 @@ type Client struct {
 	id int
 
 	key     uint32
-	counter uint32
+	counter uint16
 
 	valid bool
